% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canada_abb}
\alias{canada_abb}
\title{Postal Abbreviations for Canada}
\format{An object of class \code{character} of length 13.}
\usage{
canada_abb
}
\description{
Postal Abbreviations for Canada
}
\examples{
\dontrun{
 canada_abb
}
}
\keyword{datasets}

