% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mexico_abb}
\alias{mexico_abb}
\title{Postal Abbreviations for Mexico}
\format{An object of class \code{character} of length 32.}
\usage{
mexico_abb
}
\description{
Postal Abbreviations for Mexico
}
\examples{
\dontrun{
 mexico_abb
}
}
\keyword{datasets}

