% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{milk}
\alias{milk}
\title{Monthly milk production in Canada}
\format{A data frame with columns:
\describe{
 \item{Year}{A value between 1976 and 2015.}
 \item{Month}{A value between 1 and 12.}
 \item{Region}{Postal code abbreviation for territory or province.}
 \item{Kiloliters}{Milk sold off farms in kiloliters.}
}}
\source{
Statistics Canada. Table 003-0011 - Milk production and utilization,
monthly. \url{http://www5.statcan.gc.ca/cansim/}
}
\usage{
milk
}
\description{
Monthly milk production in Canada
}
\examples{
\dontrun{
 milk
}
}
\keyword{datasets}

