% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniusa.R
\name{miniusa}
\alias{miniusa}
\title{Make a tile grid map of The United States of America}
\usage{
miniusa(states, state_colors, border_colors = rep("white", 51),
  state_names = TRUE, state_name_colors = rep("white", 51),
  state_name_cex = 1, font = NULL)
}
\arguments{
\item{states}{A vector of US state postal abbreviations. This vector must be
some permutation of \code{usa_abb}.}

\item{state_colors}{A vector of "colors" in the R sense. For example strings
(\code{"blue"}), hex codes (\code{"#D0C7B9"}), etc. The ith color in this
vector will be the color of square that represents the ith element of
\code{states}.}

\item{border_colors}{Like \code{state_colors} but specifying the border
of the square.}

\item{state_names}{Should the postal codes for each state be displayed in the
center of the state? The default value is \code{TRUE}.}

\item{state_name_colors}{Like \code{state_colors} but specifying the color
of the text displayed in each state.}

\item{state_name_cex}{The size of the text displayed inside of each state.}

\item{font}{The font of the text displayed inside of each state. The values
\code{"serif"}, \code{"sans"}, and \code{"mono"} are safest to use. Use
other fonts at your own risk. If \code{NULL} a sans-style font will be used.}
}
\description{
Make a tile grid map of The United States of America
}
\examples{
\dontrun{
 miniusa(state_abb, 1:51)
}
}

