% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssm}
\alias{ssm}
\title{Same sex marriage in the US}
\format{A data frame with columns:
\describe{
 \item{state}{State Abbreviation}
 \item{status}{Legal status. Either \code{bbs} meaning banned by statute,
   \code{nl} meaning not legal, \code{legal}, \code{bbca} meaning banned by
   constitutional ammendment, or \code{dis} meaning disputed.}
 \item{year}{Year status went into effect.}
}}
\source{
\url{http://www.nytimes.com/interactive/2015/03/04/us/gay-marriage-state-by-state.html}
}
\usage{
ssm
}
\description{
Changes in the legality of same sex marriage in the United States over time.
}
\examples{
\dontrun{
 ssm
}
}
\keyword{datasets}

