% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa_abb}
\alias{usa_abb}
\title{Postal Abbreviations for The United States of America}
\format{An object of class \code{character} of length 51.}
\usage{
usa_abb
}
\description{
Postal Abbreviations for The United States of America
}
\examples{
\dontrun{
 usa_abb
}
}
\keyword{datasets}

