% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniusa.R
\name{miniusa}
\alias{miniusa}
\title{Make a tile grid map of the United States of America}
\usage{
miniusa(states, state_colors, border_colors = rep("white", 51),
  state_names = TRUE, state_name_colors = rep("white", 51),
  state_name_cex = 1, font = NULL)
}
\arguments{
\item{states}{description}

\item{state_colors}{description}

\item{border_colors}{description}

\item{state_names}{description}

\item{state_name_colors}{description}

\item{state_name_cex}{description}

\item{font}{description}
}
\description{
description
}
\examples{
\dontrun{
 miniusa(c(state.abb, "DC"), 1:51)
}
}

